requirements.txt
click==8.1.3
coremltools==6.3.0
mpmath==1.2.1
numpy==1.23.1
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyparsing==3.0.9
sympy==1.10.1
tqdm==4.64.0
typer==0.6.1
